// src/components/ui/Table/Table.types.ts
export interface TablePropsType {
  bordered?: boolean;
  children?: string | React.ReactNode;
  className?: string;
  hovered?: boolean;
  layout?: 'auto' | 'fixed';
  padding?: 'none' | 'normal' | 'large';
  rounded?: boolean;
  shadow?: boolean;
  size?: 'small' | 'medium' | 'large';
  transparent?: boolean;
  zebra?: boolean;
}